requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
decorator==4.4.2
distro==1.9.0
edge-tts==7.0.2
filelock==3.18.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.6.0
fsspec==2025.3.2
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
gTTS==2.5.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.31.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
langdetect==1.0.9
MarkupSafe==3.0.2
moviepy==1.0.3
multidict==6.4.3
nltk==3.9.1
numpy==2.2.5
oauthlib==3.2.2
openai==1.85.0
opencv-python-headless==4.11.0.86
packaging==25.0
pillow==11.2.1
proglog==0.1.12
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
pydub==0.25.1
pyparsing==3.2.3
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
python-dotenv==1.1.0
python-slugify==8.0.4
python-vlc==3.0.21203
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
safetensors==0.5.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.43
srt==3.5.3
stripe==12.4.0
tabulate==0.9.0
text-unidecode==1.3
tiktoken==0.9.0
tokenizers==0.21.1
tqdm==4.67.1
transformers==4.51.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
yarl==1.20.0
Flask-WTF==1.2.2
email-validator==2.2.0
flask-socketio==5.3.6
Flask-Limiter==3.5.0
